moves core XY functions to their own header
