+++ api/Makefile
@@ -40,6 +40,7 @@ libdrvapi-headers += $(DRV_DIR)/api/DrvAPIGlobal.hpp
 libdrvapi-headers += $(DRV_DIR)/api/DrvAPIOp.hpp
 libdrvapi-headers += $(DRV_DIR)/api/DrvAPISysConfig.hpp
 libdrvapi-headers += $(DRV_DIR)/api/DrvAPIInfo.hpp
+libdrvapi-headers += $(DRV_DIR)/api/DrvAPICoreXY.hpp
 libdrvapi-headers += $(DRV_DIR)/api/DrvAPIAddressToNative.hpp
 libdrvapi-headers += $(DRV_DIR)/api/DrvAPINativeToAddress.hpp
 libdrvapi-headers += $(DRV_DIR)/api/DrvAPISystem.hpp
